### 2022-04-29 #533 @nathansam
datefixR: Fix Really Messy Dates
Submitting Author Name: Nathan Constantine-Cooke
Submitting Author Github Handle: @nathansam
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/nathansam/datefixR
Version submitted: 0.1.4.9000
Submission type: Standard
Version accepted: TBD
---

```
Package: datefixR
Title: Fix Really Messy Dates
Version: 0.1.4.9000
Authors@R: person("Nathan",
                  "Constantine-Cooke",
                  email = "nathan.constantine-cooke@ed.ac.uk",
Description: Fixes messy dates in data frames such as those entered via text
  boxes. Standardizes / - and whitespace separation, month abbreviations, and
  year first or day first by converting to R's built-in Date class. Imputes
  missing date or month using user-provided values.
    lifecycle,
    stringr
```

datefixR takes date data in which has been stored in many different formats (01/01/2001, 5 April 2020, Dec 2015 etc.) and converts them to R's `Date` type.

Any researchers using data entered via a questionnaire which (unfortunately) asked for a date as free-text. Given the nature of this data generation, this mainly affects those who work with human subjects.

[`lubridate::guess_formats()`](https://lubridate.tidyverse.org/reference/guess_formats.html) can be used to guess a date format and [`lubridate::parse_date_time()`](https://lubridate.tidyverse.org/reference/parse_date_time.html) calls this function when it attempts to parse a vector into a POSIXct date-time object. However:

1.  When a date fails to parse in `{lubridate}` then the user is simply told how many dates failed to parse. In `datefixR` the user is told the ID (assumed to be the first column by default but can be user-specified) corresponding to the date which failed to parse and reports the considered date: making it much easier to figure out which dates supplied failed to parse and why.
2.  When imputing a missing day or month, there is no user-control over  this behaviour. For example, when imputing a missing month, the user may wish to impute July, the middle of the year, instead of January.  However, January will always be imputed in `{lubridate}`. In `datefixR`, this behaviour can be controlled by the `month.impute` argument.
3.  These functions require all possible date formats to be specified in the `orders` argument, which may result in a date format not being considered if the user forgets to list one of the possible formats. By contrast, `datefixR` only needs a format to be specified if month-first is to be preferred over day-first when guessing a date.

However, `{lubridate}` of course excels in general date manipulation and is an excellent tool to use alongside `datefixR`.

An alternative function is [`anytime::anydate()`](https://dirk.eddelbuettel.com/code/anytime.html) which also attempts to convert dates to a consistent format (POSIXct). However `{anytime}` assumes year, month, and day have all been provided and does not permit imputation. Moreover, if a date cannot be parsed, then the date is converted to an NA object and no warning is raised- which may lead to issues later in the analysis.

NA

https://github.com/ropensci/software-review/issues/529

pkgcheck claims: "Package has no continuous integration checks." however [Github actions is already used](https://github.com/nathansam/datefixR/actions) so this is a false positive.

(Already available on CRAN)

NA
