### 2022-05-03 f968227
Update actions to lint files
