### 2022-05-03 821ec79
Don't fail if lint not needed
