+++ .github/workflows/R-CMD-check.yaml
@@ -9,7 +9,7 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - uses: r-lib/actions/setup-r@v1
         with:
           use-public-rspm: true
@@ -20,14 +20,16 @@ jobs:
 
       - name: Lint package
         run: styler::style_pkg()
-        shell: Rscript {0}a
+        shell: Rscript {0}
       - name: Commit and push changes
         run: |
-          git config user.name 'github-actions[bot]'
-          git config user.email 'github-actions[bot]@users.noreply.github.com'
-          git add -A
-          git commit -m "Lint package"
-          git push
+          if [[ `git status --porcelain` ]]; then
+           git config user.name 'github-actions[bot]'
+           git config user.email 'github-actions[bot]@users.noreply.github.com'
+           git add -A
+           git commit -m "Lint package"
+           git push
+          fi
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
 
