### 2022-05-03 9f83dd9
Update docs to cover code style
