+++ NEWS.md
@@ -1,6 +1,7 @@
 # datefixR 0.1.5
 
 * Dates with a "sept" / "Sept" abbreviation will now be correctly handled.
+* Lint all R files using styler
 
 # datefixR 0.1.4
 
