+++ README.Rmd
@@ -127,7 +127,7 @@ date-time object. However:
   the date which failed to parse and reports the considered date: making it much
   easier to figure out which dates supplied failed to parse and why. 
 2.  When imputing a missing day or month, there is no user-control over this
-  behaviour. For example, when imputing a missing month, the user may wish to
+  behavior. For example, when imputing a missing month, the user may wish to
   impute July,  the middle of the year, instead of January. However, January will
   always be imputed in `{lubridate}`. In `datefixR`, this behaviour can be
   controlled by the `month.impute` argument.
@@ -159,10 +159,10 @@ consistently being over 3 times slower (testing up to 10,000 dates).
 of time faster than `datefixR` and `lubridate::parse_date_time2()`, which is
 written in C but only allows numeric dates, is even faster. If you are don't mind
 not having control over imputation, do not expect to have to deal with many
-unparsed dates, are confident you will specify all potential formats the
-supplied dates will be in, and you have many many dates to standardise (hundreds
-of thousands or more), `{lubridate}`'s functions may be a better option than
-`datefixR`. 
+dates which fail to parse, are confident you will specify all potential formats
+the supplied dates will be in, and you have many many dates to standardize
+(hundreds of thousands or more), `{lubridate}`'s functions may be a better
+option than `datefixR`. 
 
 ### Not actively maintained alternatives
 
