### 2022-05-03 96dd6fe
Fix lintr warnings
