+++ R/fix_dates.R
@@ -27,24 +27,30 @@ months$full <- tolower(month.name)
 #' @param format Character. The format which a date is mostly likely to be given
 #'   in. Either \code{"dmy"} (default) or \code{"mdy"}. If year appears to have
 #'   been given first, then YMD is assumed for the subject (format argument is
-#'   not used for these observations) 
+#'   not used for these observations)
 #' @return A \code{dataframe} or \code{tibble} object. Dependent on the type of
 #'   \code{df}. Selected columns are of type \code{Date}
 #' @seealso \code{\link{fix_date}} Similar to \code{fix_dates()} except can only be
-#' applied to character objects. 
+#' applied to character objects.
 #' @examples
-#' bad.dates <- data.frame(id = seq(5),
-#'                         some.dates = c("02/05/92",
-#'                                        "01-04-2020",
-#'                                        "1996/05/01",
-#'                                        "2020-05-01",
-#'                                        "02-04-96"),
-#'                         some.more.dates = c("2015",
-#'                                             "02/05/00",
-#'                                             "05/1990",
-#'                                             "2012-08",
-#'                                             "jan 2020"))
-#'fixed.df <- fix_dates(bad.dates, c("some.dates", "some.more.dates"))
+#' bad.dates <- data.frame(
+#'   id = seq(5),
+#'   some.dates = c(
+#'     "02/05/92",
+#'     "01-04-2020",
+#'     "1996/05/01",
+#'     "2020-05-01",
+#'     "02-04-96"
+#'   ),
+#'   some.more.dates = c(
+#'     "2015",
+#'     "02/05/00",
+#'     "05/1990",
+#'     "2012-08",
+#'     "jan 2020"
+#'   )
+#' )
+#' fixed.df <- fix_dates(bad.dates, c("some.dates", "some.more.dates"))
 #' @export
 fix_dates <- function(df,
                       col.names,
@@ -59,44 +65,45 @@ fix_dates <- function(df,
     stop("col.names should be a character vector!")
   }
   .checkformat(format)
-  
+
   if (is.null(id)) id <- 1 # Use first column as id if not explicitly given
 
   .checkday(day.impute)
   .checkmonth(month.impute)
   day.impute <- .convertimpute(day.impute)
   month.impute <- .convertimpute(month.impute)
-  error.status <- 0
-  
+
   for (col.name in col.names) {
     fixed.dates <- c()
-      for (i in 1:nrow(df)) {
-        tryCatch(
-          {
-            fixed.dates[i] <- .fix_date(df[i, col.name],
-                                       day.impute,
-                                       month.impute,
-                                       subject = df[i, id],
-                                       format = format)
-            },
-          error = function(cond){
-            message(paste0("Unable to resolve date for subject ",
-                       df[i, id],
-                       " (date: ",
-                       df[i, col.name],
-                       ")\n")
-                 )
-            stop(cond)}
-        )
-      }
-    df[, col.name] <- as.Date(fixed.dates)
+    for (i in seq_len(df)) {
+      tryCatch(
+        {
+          fixed.dates[i] <- .fix_date(df[i, col.name],
+            day.impute,
+            month.impute,
+            subject = df[i, id],
+            format = format
+          )
+        },
+        error = function(cond) {
+          message(paste0(
[…]
