### 2022-05-03 314a56d
Fix seq_len() call
