+++ NEWS.md
@@ -1,4 +1,4 @@
-# datefixR 0.1.5
+# datefixR 0.1.6
 
 * Dates with a "sept" / "Sept" abbreviation will now be correctly handled.
 * Lint all R files using styler
