### 2022-05-04 #533 @adamhsparks
Hi @nathansam, could you address the issues _lintr_ found, `Avoid 1:nrow(...) expressions, use seq_len.` and the lines of code >80 chars, please?

While you're doing that I'll start looking for reviewers.
