### 2022-05-04 475a77f
Respond to initial ropensci editor comments
