### 2022-05-04 #533 @nathansam
Hi @adamhsparks, thank you! Those lintr issues should now be fixed.
