+++ .github/workflows/R-CMD-check.yaml
@@ -89,11 +89,16 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: pkgdown
+          extra-packages: any::pkgdown, local::.
           needs: website
 
       - name: Deploy package
+        if: github.event_name != 'pull_request'
         run: |
           git config --local user.name "$GITHUB_ACTOR"
           git config --local user.email "$GITHUB_ACTOR@users.noreply.github.com"
           Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE)'
+      - name: Build site without deploying
+        if: github.event_name == 'pull_request'
+        run: |
+          Rscript -e 'pkgdown::build_site(preview = FALSE, install = FALSE)'
