+++ .github/workflows/R-CMD-check.yaml
@@ -58,7 +58,7 @@ jobs:
           extra-packages: any::rcmdcheck
           needs: check
       - name: Check
-        run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "note", check_dir = "check")
+        run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
       - name: Install dependencies
         run: |
