### 2022-05-09 #533 @adamhsparks
Hi @nathansam, as I've been looking for reviewers and checking over the package some more, I did note some issues with the spelling and tests as well.

When I check the tests, I get this:

```r
Backtrace:
 1. testthat::expect_equal(...)
      at test_fix_dates.R:206:2
 4. datefixR::fix_dates(temp, "date", day.impute = NA)
 9. datefixR::.fix_date(...)
      at datefixR/R/fix_dates.R:81:10

── Warning (test_fix_dates.R:215:3): Imputing month with NA results in NA ──────
NA imputed for subject 1 (date: 1994)

Backtrace:
 1. testthat::expect_equal(...)
      at test_fix_dates.R:215:2
 4. datefixR::fix_dates(temp, "date", month.impute = NA)
 9. datefixR::.fix_date(...)
      at datefixR/R/fix_dates.R:81:10
```

With the spelling, your DESCRIPTION says "en-US" but you've used British English for "behaviour", also the spellcheck is picking up a few other words that are non-standard but probably correct anyway. I suggest setting up spell checking with `usethis::use_spell_check()`.
