### 2022-05-09 5914733
Fix typos and add check for spelling
ropensci/software-review#533
