### 2022-05-09 4436d27
Add explicit tests for warnings
ropensci/software-review#533
