+++ tests/testthat/test_fix_dates.R
@@ -203,17 +203,16 @@ test_that("Error if day out of bounds", {
 
 test_that("Imputing day with NA results in NA", {
   temp <- data.frame(id = 1, date = "04-1994")
-  expect_equal(
-    fix_dates(temp, "date", day.impute = NA),
-    data.frame(id = 1, date = as.Date(NA))
-  )
+  expect_warning(date.guess <- fix_dates(temp, "date", day.impute = NA))
+  expect_equal(date.guess, data.frame(id = 1, date = as.Date(NA)))
 })
 
 
 test_that("Imputing month with NA results in NA", {
   temp <- data.frame(id = 1, date = "1994")
+  expect_warning(date.guess <- fix_dates(temp, "date", month.impute = NA))
   expect_equal(
-    fix_dates(temp, "date", month.impute = NA),
+    date.guess,
     data.frame(id = 1, date = as.Date(NA))
   )
 })
