### 2022-05-09 ef7f3fb
Update NEWS.md
