### 2022-05-09 #533 @nathansam
Thanks @adamhsparks ! I have fixed the typo and have now set up `{spelling}` as you have suggested. Thanks for the suggestion; I will definitely be using it across all R packages I maintain. Thank you for catching the warnings in the tests. The warnings are intended (but were added after the tests were first written), and I have now added explicit `expect_warning()` statements to those tests.
