### 2022-05-09 6f807b9
Brush up on my regex
ropensci/software-review/533
