+++ .github/workflows/R-CMD-check.yaml
@@ -32,7 +32,7 @@ jobs:
            git config user.name 'github-actions[bot]'
            git config user.email 'github-actions[bot]@users.noreply.github.com'
            git add -A
-           git commit -m "Lint package"
+           git commit -m "Re-render README"
            git push
           fi
           
