### 2022-05-09 840ad27
Use pak to install package + deps
