+++ .github/workflows/R-CMD-check.yaml
@@ -42,11 +42,11 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
-
-      - name: Install styler
-        run: install.packages(c("styler", "roxygen2"))
-        shell: Rscript {0}
-
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          extra-packages: | 
+            any::styler
+            any::roxygen2
       - name: Lint package
         run: styler::style_pkg()
         shell: Rscript {0}
