+++ .github/workflows/R-CMD-check.yaml
@@ -47,6 +47,7 @@ jobs:
           extra-packages: | 
             any::styler
             any::roxygen2
+          dependencies: false
       - name: Lint package
         run: styler::style_pkg()
         shell: Rscript {0}
