+++ .github/workflows/R-CMD-check.yaml
@@ -47,7 +47,6 @@ jobs:
           extra-packages: | 
             any::styler
             any::roxygen2
-          dependencies: "FALSE"
       - name: Lint package
         run: styler::style_pkg()
         shell: Rscript {0}
