+++ R/fix_dates.R
@@ -165,10 +165,7 @@ fix_dates <- function(df,
       }
     }
   }
-
-  #' @noRd
   .checkoutput(day, month)
-
   if (is.na(day) || is.na(month)) {
     fixed_date <- NA
     warning(paste0(
