### 2022-05-12 e8fbc89
Reduce cyclomatic complexity
