+++ R/internal.R
@@ -1,0 +1,239 @@
+
+#' @noRd
+.fix_date <- function(date,
+                      day.impute,
+                      month.impute,
+                      subject,
+                      format = format) {
+  if (is.null(date) || is.na(date) || as.character(date) == "") {
+    return(NA)
+  }
+  date <- as.character(date)
+  date <- .convert_text_month(date)
+  
+  if (nchar(date) == 4) {
+    # Just given year
+    year <- date
+    month <- .imputemonth(month.impute)
+    day <- .imputeday(day.impute)
+  } else {
+    date_vec <- .separate_date(date)
+    if (any(nchar(date_vec) > 4)) {
+      stop("unable to tidy a date")
+    }
+    date_vec <- .appendyear(date_vec)
+    
+    if (length(date_vec) < 3) {
+      # ASSUME MM/YYYY, YYYY/MM
+      day <- .imputeday(day.impute)
+      if (nchar(date_vec[1]) == 4) {
+        # Assume YYYY/MM
+        year <- date_vec[1]
+        month <- date_vec[2]
+      } else if (nchar(date_vec[2]) == 4) {
+        year <- date_vec[2]
+        month <- date_vec[1]
+      }
+    } else {
+      if (nchar(date_vec[1]) == 4) {
+        # Assume YYYY/MM/DD
+        year <- date_vec[1]
+        month <- date_vec[2]
+        day <- date_vec[3]
+      } else {
+        if (format == "dmy") {
+          # Assume DD/MM/YYYY
+          year <- date_vec[3]
+          month <- date_vec[2]
+          day <- date_vec[1]
+        }
+        if (format == "mdy") {
+          year <- date_vec[3]
+          month <- date_vec[1]
+          day <- date_vec[2]
+        }
+      }
+    }
+  }
+  .checkoutput(day, month)
+  .combinepartialdate(day, month, year, date, subject)
+}
+
+#' @noRd
+.separate_date <- function(date) {
+  if (grepl("/", date, fixed = TRUE)) {
+    date_vec <- stringr::str_split_fixed(date,
+                                         pattern = "/",
+                                         n = Inf
+    )
+  } else if (grepl("-", date, fixed = TRUE)) {
+    date_vec <- stringr::str_split_fixed(date,
+                                         pattern = "-",
+                                         n = Inf
+    )
+  } else if (grepl(" ", date, fixed = TRUE)) {
+    date_vec <- stringr::str_split_fixed(date,
+                                         pattern = " ",
+                                         n = Inf
+    )
+  }
+  date_vec
+}
+
+#' @noRd
+.convert_text_month <- function(date) {
+  date <- tolower(date)
+  # Handle special case of sept abbrv.
+  date <- gsub(
+    pattern = "sept",
+    replacement = "09",
+    x = date
+  )
+  
+  for (i in 1:12) {
+    if (i < 10) {
+      replacement <- paste0("0", i)
+    } else {
+      replacement <- i
+    }
+    date <- gsub(
+      pattern = months$full[i],
+      replacement = replacement,
+      x = date
+    )
+    
+    date <- gsub(
+      pattern = months$abbrev[i],
+      replacement = replacement,
+      x = date
+    )
+  }
+  date
+}
+
+#' @noRd
+.checkday <- function(day.impute) {
+  if (!is.na(day.impute) && !is.null(day.impute)) {
+    if (day.impute < 1 | day.impute > 28) {
+      stop("day.impute should be an integer between 1 and 28\n")
+    }
[…]
