### 2022-05-12 978f67d
Remove references to datefixR not supporting mdy formats
