### 2022-05-31 448a154
Update usage doc in README
