### 2022-06-02 #533 @arhepworth
Apologies, have dropped the ball on this one. I can theoretically get it done by the end of the weekend, if that is useful?

From: Nathan Constantine-Cooke ***@***.***>
Date: Wednesday, 1 June 2022 at 20:15
To: ropensci/software-review ***@***.***>
Cc: Anna Hepworth ***@***.***>, Mention ***@***.***>
Subject: [EXT] - Re: [ropensci/software-review] datefixR: Fix Really Messy Dates (Issue #533)

CAUTION: This email originated from outside of DPIRD. Do not click links or open attachments unless you recognise the sender and know the content is safe.

Hi everyone! I hope you are all well! Just dropping a reminder the review due date was yesterday. Thanks for agreeing to review datefixR 🙂.
