### 2022-06-02 #533 @AlvesKS
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

The Package `datefixR` developed by Constantine-Cooke provides two functions to convert dates into the R standardized format (`ymd`). Besides the functions cannot distinguish between `dmy` and `mdy` (US format) date types when both are present in the same column, they work pretty well for what they claim to do.

Below, I detail some issues and recommendations for what I think would improve the package.

I recommend stating clearly what is the purpose of the package in the README file. The way it is stated in the first paragraph of the README file is okay, but I think it could be more detailed.

#### Function naming

I think the names of the function can be easily confused. The authors should consider changing the names to `fixdate_df()` and `fixdate_char()`.

#### Interaction with `dplyr::mutate()`

I notice that it is not possible to use `fix_date()` in the `mutate()` function from `dplyr` package. I recommend adding this functionality.

#### `fix_dates()` documentation

In the `fix_dates()` description (documentation), it is stated "Cleans up a `dataframe` object which...". Consider replacing "Cleans up" with "correct", "fix up", "tidy", "adjust", etc. When I first read "Cleans up", I  thought that the function would erase the data from the data frame.

#### Vignette

Provide a function to call the first data frame (`bad.dates`) provided in the package Vignette. It would be faster to reproduce the examples. After looking at the function documentation, I saw that the code to build the `bad.dates` data frame (with different data) is provided in the `fix_dates()` documentation (help material), but I think having a ready to use data frame is better. Another option is providing the code to build the data frame in the Vignette.
