### 2022-06-06 #533 @nathansam
Thank you for your feedback @AlvesKS ! It is all very much appreciated and is a big help. I will begin implementing your suggestions on an `ropensci` branch of the datefixR repo so changes aren't being pushed to main whilst @arhepworth is still reviewing.
