### 2022-06-06 2f60221
- change function names
fix_dates and fix_date are now fix_date_df and fix_date respectively
