+++ R/datefixR-package.R
@@ -1,0 +1,23 @@
+#' datefixR: Fix Really Messy Dates
+#'
+#' @section Description:
+#' Fixes messy dates in data frames such as those entered via text boxes.
+#' Standardizes / - and whitespace separation, month abbreviations, and year
+#' first or day first by converting to R's built-in Date class. Imputes missing
+#' date or month using user-provided values.
+#'
+#' @section Author:
+#' \strong{Maintainer}: Nathan Constantine-Cooke
+#' \email{nathan.constantine-cooke@@ed.ac.uk}
+#' \href{https://orcid.org/0000-0002-4437-8713}{(ORCID)}
+#'
+#' @seealso
+#' \url{https://www.constantine-cooke.com/datefixR/}
+#'
+#' \url{https://www.github.com/nathansam/datefixR/}
+## usethis namespace: start
+#' @importFrom lifecycle deprecated
+## usethis namespace: end
+#' @docType package
+#' @name datefixR
+NULL
