+++ R/datefixR-package.R
@@ -1,10 +1,9 @@
 #' datefixR: Fix Really Messy Dates
 #'
-#' @section Description:
-#' Fixes messy dates in data frames such as those entered via text boxes.
-#' Standardizes / - and whitespace separation, month abbreviations, and year
-#' first or day first by converting to R's built-in Date class. Imputes missing
-#' date or month using user-provided values.
+#' @description Fixes messy dates in data frames such as those entered via text
+#' boxes. Standardizes / - and whitespace separation, month abbreviations, and
+#' year first or day first by converting to R's built-in Date class. Imputes
+#' missing date or month using user-provided values.
 #'
 #' @section Author:
 #' \strong{Maintainer}: Nathan Constantine-Cooke
@@ -14,7 +13,7 @@
 #' @seealso
 #' \url{https://www.constantine-cooke.com/datefixR/}
 #'
-#' \url{https://www.github.com/nathansam/datefixR/}
+#' \url{https://github.com/nathansam/datefixR/}
 ## usethis namespace: start
 #' @importFrom lifecycle deprecated
 ## usethis namespace: end
