+++ R/fix_date_df.R
@@ -3,7 +3,7 @@ months$abbrev <- tolower(month.abb)
 months$full <- tolower(month.name)
 
 #' @title Clean up messy date columns
-#' @description Cleans up a \code{dataframe} object which has date columns
+#' @description Tidies a \code{dataframe} object which has date columns
 #' entered via a free-text box (possibly by different users) and are therefore
 #' in a non-standardized format. Supports numerous separators including /,-, or
 #' space. Supports all-numeric, abbreviation, or long-hand month notation. Where
