### 2022-06-09 b6ea766
- Vectorise fix_date_char
