### 2022-06-09 d31424d
Vectorise fix_date_char
