### 2022-06-10 8219ac0
Update tests to handle deprecated funcs
