+++ tests/testthat/_snaps/fix_date.md
@@ -1,0 +1,10 @@
+# fix_date is deprecated
+
+    Code
+      fix_date(bad.date)
+    Warning <lifecycle_warning_deprecated>
+      `fix_date()` was deprecated in datefixR 1.0.0.
+      Please use `fix_date_char()` instead.
+    Output
+      [1] "2021-03-02"
+
