+++ DESCRIPTION
@@ -25,7 +25,8 @@ Suggests:
     rmarkdown,
     knitr,
     testthat (>= 3.0.0),
-    spelling
+    spelling,
+    withr
 Config/testthat/edition: 3
 VignetteBuilder: knitr
 LazyData: true
