### 2022-06-10 8e856b8
Give date in warning for fix_date_char
Now the function applies to vectors, it makes sense to provide the user with name of dates which cause a warning
