### 2022-06-10 1ce761b
Update documentation
Reflects change to supporting character vectors and remove references to deprecated functions
