+++ R/fix_date_char.R
@@ -1,22 +1,22 @@
 #' @title Convert improperly formatted date to R's Date class
 #' @description
-#' Converts a single improperly formatted date to R's Date class.
-#' Supports numerous separators including /,- or white space.
-#' Supports all-numeric, abbreviation or long-hand month notation. Where
-#' day of the month has not been supplied, the first day of the month is
-#' imputed. Either DMY or YMD is assumed by default. However, the US system of
-#' MDY is supported via the \code{format} argument.
-#' @param date Character to be converted to R's date class.
+#' Converts a character vector (or single character object)  from improperly
+#' formatted dates to R's Date class. Supports numerous separators including /,-
+#' or white space. Supports all-numeric, abbreviation or long-hand month
+#' notation. Where day of the month has not been supplied, the first day of the
+#' month is imputed by default. Either DMY or YMD is assumed by default.
+#' However, the US system of MDY is supported via the \code{format} argument.
+#' @param dates Character vector to be converted to R's date class.
 #' @inheritParams fix_date_df
-#' @return An object belonging to R's built in \code{Date} class.
-#' @seealso \code{\link{fix_dates}} Similar to \code{fix_date()} except is
+#' @return A vector of elements belonging to R's built in \code{Date} class.
+#' @seealso \code{\link{fix_date_df}} Similar to \code{fix_date()} except is
 #' applicable to columns of a dataframe.
 #' @examples
 #' bad.date <- "02 03 2021"
 #' fixed.date <- fix_date_char(bad.date)
 #' fixed.date
 #' @export
-fix_date_char <- function(date,
+fix_date_char <- function(dates,
                           day.impute = 1,
                           month.impute = 7,
                           format = "dmy") {
@@ -24,7 +24,7 @@ fix_date_char <- function(date,
   .checkmonth(month.impute)
   .checkformat(format)
 
-  as.Date(sapply(date,
+  as.Date(sapply(dates,
     .fix_date_char,
     day.impute = day.impute,
     month.impute = month.impute,
