+++ vignettes/datefixR.Rmd
@@ -22,8 +22,8 @@ This vignette describes the functionality of `datefixR` in more detail than the
 [README](https://github.com/nathansam/datefixR). `DatefixR` is a lightweight
 package consisting of a two user-accessible functions, `fix_date_char()` and
 `fix_date_df()`, which converts dates written in different formats into R's
-built-in `Date` class. The former is designed to modify single character objects
-whilst the latter is intended for modifying columns of a data frame (or tibble).
+built-in `Date` class. The former is designed to modify character vectors whilst
+the latter is intended for modifying columns of a data frame (or tibble).
 
 Practically, this package is most useful when handling date data which has been
 supplied via text boxes (instead of a date-specific input with a consistent date
@@ -39,33 +39,41 @@ consider a data frame with two columns of dates in differing formats
 with no missing data. 
 
 ```{R, echo = FALSE}
-bad.dates <- data.frame(id = seq(5),
-                        some.dates = c("02/05/92",
-                                       "01-04-2020",
-                                       "1996/05/01",
-                                       "2020-05-01",
-                                       "02-04-96"),
-                        some.more.dates = c("01 03 2015",
-                                            "02/05/00",
-                                            "01/05/1990",
-                                            "03-Dec-2012",
-                                            "02 April 2020"))
+bad.dates <- data.frame(
+  id = seq(5),
+  some.dates = c(
+    "02/05/92",
+    "01-04-2020",
+    "1996/05/01",
+    "2020-05-01",
+    "02-04-96"
+  ),
+  some.more.dates = c(
+    "01 03 2015",
+    "02/05/00",
+    "01/05/1990",
+    "03-Dec-2012",
+    "02 April 2020"
+  )
+)
 knitr::kable(bad.dates)
 ```
 
-`fix_date_df()` requires two arguments, `df`, a data frame (or tibble) object and
-`col.names`, a character vector containing the names of columns to be
-standardized. By default, the first column of the data frame is assumed to contain
-row IDs. These IDs are used if a warning or error is raised to assist with
-locating the source of the error. The ID column can also be manually provided
-via the `id` argument. 
+`fix_date_df()` requires two arguments, `df`, a data frame (or tibble) object
+and `col.names`, a character vector containing the names of columns to be
+standardized. By default, the first column of the data frame is assumed to
+contain row IDs. These IDs are used if a warning or error is raised to assist
+with locating the source of the error. The ID column can also be manually
+provided via the `id` argument. 
 
 The output from this function is a data frame object with the selected date
 columns now belonging to the `Date` class. 
 
 ```{R}
-fixed.dates <- fix_date_df(bad.dates,
-                         c("some.dates", "some.more.dates"))
+fixed.dates <- fix_date_df(
+  bad.dates,
+  c("some.dates", "some.more.dates")
+)
 knitr::kable(fixed.dates)
 ```
 
@@ -93,19 +101,10 @@ fix_date_char("01 02 2014", format = "mdy")
 ### Date and month imputation
 
 By default, `datefixR` imputes missing months as July, and missing days of the
-month as the first day.  As such, 
-
-```{R, echo = FALSE}
-impute.dates <- data.frame(id = 1,
-                        some.date = "1992")
-knitr::kable(impute.dates)
-```
-
-converts to 
+month as the first day.  As such, "1992" converts to 
 
 ```{R}
-imputed.dates <- fix_date_df(impute.dates, "some.date")
-knitr::kable(imputed.dates)
+fix_date_char("1992")
 ```
 
[…]
