+++ README.Rmd
@@ -25,13 +25,18 @@ knitr::opts_chunk$set(
 | | [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/533_status.svg)](https://github.com/ropensci/software-review/issues/533) | [![Tidyverse style guide](https://img.shields.io/static/v1?label=Code%20Style&message=Tidyverse&color=1f1c30)](https://style.tidyverse.org) |
 <!-- badges: end -->
 
-`datefixR` is designed to standardize messy date data, such as dates
-entered by different people via text boxes, by converting the dates to R's Date
-data type.
+`datefixR` standardizes dates in different formats or with missing data: for
+example dates which have been provided from free text web forms. 
 
-This package arose from my own fights with messy date data where dates were
-written in many different formats e.g 01-jan-15, 2015 04 02, 10/12/2010 and
-more. 
+There are many different formats dates are commonly represented with: the
+order of day, month, or year can differ, different separators ("-", "/", or
+whitespace) can be used, months can be numerical, names, or abbreviations and
+year given as two digits or four. `datefixR` takes dates in all these different
+formats and converts them to R's built-in date class. If `datefixR` cannot
+standardize a date, such as because it is too malformed, then the user is told
+which date cannot be standardized and the corresponding ID for the row.
+`datefixR` also allows the imputation of missing days and months with
+user-controlled behavior. 
 
 ## Installation instructions
 
