### 2022-07-06 nathansam/datefixR#17 @alstat
Docs: add format output in the value returned by `fix_date`
Maybe it would be helpful if we extend the statement on the Value returned by `fix_date` to the following:

> An object belonging to R's built in Date class with the following format `yyyy-mm-dd`

Just to easily understand the format of the output, although this can be seen if we manually map the input to output and see where the date, month and year fall based on the input `format` argument of `fix_date`, but I think it would be easy to understand if we have it stated already in the value returned in the docs.
