### 2022-07-06 nathansam/datefixR#18 @alstat
Docs: add datetime input as limitation
Dates is often tied up with time, and I often assume that date libraries also caters datetime input with it, which is out of the scope of this package (upon testing it), wouldn't it be useful therefore to add this to the limitations of the package that datefixR does not include date time inputs, e.g. "2021-05-01 08:20:30"? Although, it is clear from the name already and the statement on the purpose of the package, I still assumed it actually, and maybe that's just me, so you can also ignore this if you think it is well stated.
