### 2022-07-06 #533 @alstat
Hi @nathansam, I have some suggestions on [datefixR#17](https://github.com/nathansam/datefixR/issues/17), [datefixR#18](https://github.com/nathansam/datefixR/issues/18).
