### 2022-07-06 3e6e55d
Give returned date format in docs
closes #17
