### 2022-07-06 d7e05df
List lack of support for datetime as limitation
Closes #18
