### 2022-07-07 #533 @adamhsparks
Hi @alstat, please don't forget to complete our [review template](https://devguide.ropensci.org/reviewtemplate.html#reviewtemplate) along with the issues that you've opened for datefixR when you're finished with your review. 🙏
