### 2022-07-07 #533 @alstat
### Issues opened
- [ ] Docs: add format output in the value returned by fix_date [#17](https://github.com/nathansam/datefixR/issues/17)
- [ ] Docs: add datetime input as limitation [#18](https://github.com/nathansam/datefixR/issues/18)
---
