### 2022-07-07 nathansam/datefixR#18 @alstat
Addressed by [d7e05df](https://github.com/nathansam/datefixR/commit/d7e05df3a1bab4b39a35f484fc639da8db95379f)
