### 2022-07-16 #533 @alstat
I'm satisfied with the changes [here](https://github.com/nathansam/datefixR/compare/main...ropensci), both issues I opened were addressed.
