### 2022-07-16 e6bc39a
- Use vapply instead of sapply
Closes #19
