### 2022-07-16 b5b2715
Use \R{} in documentation
Closes #21
