+++ R/fix_date.R
@@ -2,15 +2,15 @@
 #' @description
 #' `r lifecycle::badge("deprecated")`
 #'
-#' Converts a single improperly formatted date to R's Date class.
+#' Converts a single improperly formatted date to \R{}'s Date class.
 #' Supports numerous separators including /,- or white space.
 #' Supports all-numeric, abbreviation or long-hand month notation. Where
 #' day of the month has not been supplied, the first day of the month is
 #' imputed. Either DMY or YMD is assumed by default. However, the US system of
 #' MDY is supported via the \code{format} argument.
-#' @param date Character to be converted to R's date class.
+#' @param date Character to be converted to \R{}'s date class.
 #' @inheritParams fix_dates
-#' @return An object belonging to R's built in \code{Date} class.
+#' @return An object belonging to \R's built in \code{Date} class.
 #' @seealso \code{\link{fix_dates}} Similar to \code{fix_date()} except is
 #' applicable to columns of a dataframe.
 #' @examples
