+++ man/datefixR.Rd
@@ -3,12 +3,17 @@
 \docType{package}
 \name{datefixR}
 \alias{datefixR}
-\title{datefixR: Fix Really Messy Dates}
+\title{datefixR: Standardize Dates in Different Formats or with Missing Data}
 \description{
-Fixes messy dates in data frames such as those entered via text
-boxes. Standardizes / - and whitespace separation, month abbreviations, and
-year first or day first by converting to R's built-in Date class. Imputes
-missing date or month using user-provided values.
+There are many different formats dates are commonly represented
+with: the order of day, month, or year can differ, different separators
+("-", "/", or whitespace) can be used, months can be numerical, names, or
+abbreviations and year given as two digits or four. \code{datefixR} takes dates
+in all these different formats and converts them to \R{}'s built-in date
+class. If \code{datefixR} cannot standardize a date, such as because it is too
+malformed, then the user is told which date cannot be standardized and the
+corresponding ID for the row. \code{datefixR} also allows the imputation of
+missing days and months with user-controlled behavior.
 }
 \section{Author}{
 
