+++ man/fix_date_char.Rd
@@ -2,12 +2,12 @@
 % Please edit documentation in R/fix_date_char.R
 \name{fix_date_char}
 \alias{fix_date_char}
-\title{Convert improperly formatted date to R's Date class}
+\title{Convert improperly formatted date to \R{}'s Date class}
 \usage{
 fix_date_char(dates, day.impute = 1, month.impute = 7, format = "dmy")
 }
 \arguments{
-\item{dates}{Character vector to be converted to R's date class.}
+\item{dates}{Character vector to be converted to \R{}'s date class.}
 
 \item{day.impute}{Integer. Day of the month to be imputed if not available.
 defaults to 1. If \code{day.impute = NA} then \code{NA} will be imputed for
@@ -26,13 +26,13 @@ been given first, then YMD is assumed for the subject (format argument is
 not used for these observations)}
 }
 \value{
-A vector of elements belonging to R's built in \code{Date} class with
-the following format \code{yyyy-mm-dd}.
+A vector of elements belonging to \R{}'s built in \code{Date} class
+with the following format \code{yyyy-mm-dd}.
 }
 \description{
 Converts a character vector (or single character object)  from improperly
-formatted dates to R's Date class. Supports numerous separators including /,-
-or white space. Supports all-numeric, abbreviation or long-hand month
+formatted dates to \R{}'s Date class. Supports numerous separators including
+/, -, or white space. Supports all-numeric, abbreviation or long-hand month
 notation. Where day of the month has not been supplied, the first day of the
 month is imputed by default. Either DMY or YMD is assumed by default.
 However, the US system of MDY is supported via the \code{format} argument.
