### 2022-07-16 09d7eb6
Fix missing code format for datefixR name
Closes  #23
