### 2022-07-16 #533 @ropensci-review-bot
Approved! Thanks @nathansam for submitting and @AlvesKS, @alstat for your reviews! :grin:
