### 2022-07-16 #533 @adamhsparks
Thank you, @AlvesKS and @alstat for your kind and insightful reviews. I hope you found this experience useful and enjoyable. @nathansam, thank you for submitting datefixR. You can transfer it to rOpenSci now.
