### 2022-07-16 #533 @nathansam
Thank you @AlvesKS and @alstat, your reviews were immeasurably helpful! I am of course more than happy to acknowledge you both as reviewers in DESCRIPTION. I have added your details based on the DESCRIPTION files for your own packages. Please let me know if you would like any edits made to your listings in the description.
https://github.com/nathansam/datefixR/blob/3a2b33b657c19fce085f57ac23a831f043838da8/DESCRIPTION

Thank you @adamhsparks for your editorship throughout and for the issues you have opened. I will begin the process of migrating datefixR to the ropensci organization now.
